+++ resources/assets/js/datatable/api.ts
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 
-import ctx, {init} from '../ctx';
-import {parseData} from '../util';
+import {init} from '../ctx';
+import {parseData} from '../lib/util';
 import render from './render';
 import Settings = DataTables.Settings;
 import ColumnSettings = DataTables.ColumnSettings;
@@ -28,10 +28,7 @@ declare global {
 	}
 }
 
-ctx.addHandler('.ctx-datatable', initDataTable);
-ctx.addHandler('.ctx-offline-datatable', initOfflineDataTable);
-
-async function initDataTable(el: HTMLElement) {
+export async function initDataTable(el: HTMLElement) {
 	await import(/*webpackChunkName: "bootstrap"*/'./bootstrap');
 
 	const $el = $(el);
@@ -56,7 +53,7 @@ async function initDataTable(el: HTMLElement) {
 	table.on('childRow.dt', (event, data) => init(data.container.get(0)));
 }
 
-async function initOfflineDataTable(el: HTMLElement) {
+export async function initOfflineDataTable(el: HTMLElement) {
 	await import(/*webpackChunkName: "bootstrap"*/'./bootstrap');
 
 	$(el).DataTable(parseData(el));
