+++ resources/assets/js/formulier.ts
@@ -1,25 +1,12 @@
 import axios from 'axios';
 import $ from 'jquery';
-import {ajaxRequest} from './ajax';
-import {bbCodeSet} from './bbcode-set';
-import {domUpdate} from './context';
-import ctx from './ctx';
 import {DatatableResponse, fnGetSelection, fnUpdateDataTable} from './datatable/api';
+import {ajaxRequest} from './lib/ajax';
 
-import {modalClose} from './modal';
+import {modalClose} from './lib/modal';
 
-import {redirect, reload} from './reload';
-
-ctx.addHandlers({
-	'.InlineFormToggle': (el) => el.addEventListener('click', (event) => formToggle(el, event)),
-	'.SubmitChange': (el) => el.addEventListener('change', formSubmit),
-	'.cancel': (el) => el.addEventListener('click', formCancel),
-	'.reset': (el) => el.addEventListener('click', formReset),
-	'.submit': (el) => el.addEventListener('click', formSubmit),
-	'form.Formulier': (el) => $(el).on('submit', formSubmit), // dit is sterker dan addEventListener
-	'textarea.BBCodeField': (el) => $(el).markItUp(bbCodeSet),
-	'time.timeago': (el) => $(el).timeago(),
-});
+import {domUpdate} from './lib/domUpdate';
+import {redirect, reload} from './lib/reload';
 
 export function formIsChanged(form: JQuery<EventTarget>) {
 	let changed = false;
