+++ resources/assets/js/lib/ajax.ts
