+++ resources/assets/js/lib/bbcode-hints.ts
@@ -1,8 +1,4 @@
-import ctx from './ctx';
-
 const {Textcomplete, Textarea} = require('textcomplete');
-
-ctx.addHandler('textarea.BBCodeField', activeerLidHints);
 
 export function activeerLidHints(textarea: HTMLElement) {
 	const editor = new Textarea(textarea);
