+++ resources/assets/js/lib/external/jquery-ui-sliderAccess.js
