+++ resources/assets/js/lib/external/jquery.markitup.js
