+++ resources/assets/js/lib/external/three.detector.ts
