+++ resources/assets/js/lib/modal.ts
