+++ resources/assets/js/router.ts
@@ -1,11 +1,11 @@
-import {route} from './util';
+import {route} from './lib/util';
 
 /**
  * Voer specifieke code uit voor specifieke routes.
  */
 
-route('/instellingen', () => import(/* webpackChunkName: "instellingen" */ './instellingen'));
-route('/documenten', () => import(/* webpackChunkName: "documenten" */ './documenten'));
+route('/instellingen', () => import(/* webpackChunkName: "instellingen" */ './page/instellingen'));
+route('/documenten', () => import(/* webpackChunkName: "documenten" */ './page/documenten'));
 route('/fotoalbum', () => import(/* webpackChunkName: "fotoalbum" */'./fotoalbum/main'));
-route('/bibliotheek', () => import(/* webpackChunkName: "bibliotheek" */'./bibliotheek'));
-route('/agenda', () => import(/* webpackChunkName: "agenda" */'./agenda'));
+route('/bibliotheek', () => import(/* webpackChunkName: "bibliotheek" */'./page/bibliotheek'));
+route('/agenda', () => import(/* webpackChunkName: "agenda" */'./page/agenda'));
